fix: properties browser to handle dart points
Redo fix from #908 as that was accidentally uploaded to the main folder
instead of the real source code locations

Closes: #907
