+++ src/app/seamly2d/core/vtooloptionspropertybrowser.cpp
@@ -744,7 +744,7 @@ QMap<QString, quint32> VToolOptionsPropertyBrowser::getObjectList(Tool *tool, GO
     quint32 toolId = tool->getId();
     QHash<quint32, QSharedPointer<VGObject>> objects;
 
-    QVector<VToolRecord> history = qApp->getCurrentDocument()->getBlockHistory();
+    QVector<VToolRecord> history = qApp->getCurrentDocument()->getLocalHistory();
     for (qint32 i = 0; i < history.size(); ++i)
     {
         const VToolRecord record = history.at(i);
@@ -753,12 +753,31 @@ QMap<QString, quint32> VToolOptionsPropertyBrowser::getObjectList(Tool *tool, GO
         {
             switch (static_cast<int>(record.getTypeTool()))
             {
+                case 42:    //Tool::True Darts
+                {
+                    QVector<quint32> list = qApp->getCurrentDocument()->getDartItems(recId);
+                    for (qint32 j = 0; j < list.size(); ++j)
+                    {
+                        quint32 id = list.at(j);
+                            VToolRecord newRecord = VToolRecord(id, record.getTypeTool(),
+                                                        qApp->getCurrentDocument()->getActiveDraftBlockName());
+
+                        const QHash<quint32, QSharedPointer<VGObject> > *objs = m_data->DataGObjects();
+                        if (objs->contains(id)) //Avoid badId Get GObject only if not a line tool which is not an object
+                        {
+                            QSharedPointer<VGObject> obj = m_data->GetGObject(id);
+                            objects.insert(id, obj);
+                        }
+                    }
+                    break;
+                }
+
                 case 45:    //Tool::Rotation
                 case 46:    //Tool::MirrorByLine
                 case 47:    //Tool::MirrorByAxis
                 case 48:    //Tool::Move
                 {
-                    QVector<quint32> list = qApp->getCurrentDocument()->getOpItems(recId, QStringLiteral("destination"));
+                    QVector<quint32> list = qApp->getCurrentDocument()->getOpItems(recId);
                     for (qint32 j = 0; j < list.size(); ++j)
                     {
                         quint32 id = list.at(j);
